move code to new recording class (basic)
