+++ src/rec.py
@@ -32,46 +32,46 @@ gi.require_version('Gdk', '3.0')
 from gi.repository import Gdk, Gio, GLib, Gst, GstPbutils
 
 
-def on__frames_changed(self, *args):
-    frames = self.settings.get_enum("frames-per-second")
-    if frames == 0:
-        self.videoFrames = 15
-    if frames == 1:
-        self.videoFrames = 30
-    if frames == 2:
-        self.videoFrames = 60
-    return self.videoFrames
+# def on__frames_changed(self, *args):
+#     frames = self.settings.get_enum("frames-per-second")
+#     if frames == 0:
+#         self.videoFrames = 15
+#     if frames == 1:
+#         self.videoFrames = 30
+#     if frames == 2:
+#         self.videoFrames = 60
+#     return self.videoFrames
 
 
-def on__quality_changed(self, *args):
-    quality = self.settings.get_boolean("high-video-quality")
-    self.recordFormat = on__formats_changed(self, *args)
-    if quality:  # high quality
-        if self.recordFormat == "webm" or self.recordFormat == "mkv":
-            self.quality_video = "vp8enc min_quantizer=25 max_quantizer=25 cpu-used={0} cq_level=13 deadline=1000000 threads={0}".format(
-                self.cpus)
-        elif self.recordFormat == "mp4":
-            self.quality_video = "x264enc qp-min=17 qp-max=17 speed-preset=1 threads={0} ! h264parse ! video/x-h264, profile=baseline".format(
-                self.cpus)
-    else:
-        if self.recordFormat == "webm" or self.recordFormat == "mkv":
-            self.quality_video = "vp8enc min_quantizer=5 max_quantizer=10 cpu-used={0} cq_level=13 deadline=1000000 threads={0}".format(
-                self.cpus)
-        elif self.recordFormat == "mp4":
-            self.quality_video = "x264enc qp-min=5 qp-max=5 speed-preset=1 threads={0} ! h264parse ! video/x-h264, profile=baseline".format(
-                self.cpus)
-    return self.quality_video
+# def on__quality_changed(self, *args):
+#     quality = self.settings.get_boolean("high-video-quality")
+#     self.recordFormat = on__formats_changed(self, *args)
+#     if quality:  # high quality
+#         if self.recordFormat == "webm" or self.recordFormat == "mkv":
+#             self.quality_video = "vp8enc min_quantizer=25 max_quantizer=25 cpu-used={0} cq_level=13 deadline=1000000 threads={0}".format(
+#                 self.cpus)
+#         elif self.recordFormat == "mp4":
+#             self.quality_video = "x264enc qp-min=17 qp-max=17 speed-preset=1 threads={0} ! h264parse ! video/x-h264, profile=baseline".format(
+#                 self.cpus)
+#     else:
+#         if self.recordFormat == "webm" or self.recordFormat == "mkv":
+#             self.quality_video = "vp8enc min_quantizer=5 max_quantizer=10 cpu-used={0} cq_level=13 deadline=1000000 threads={0}".format(
+#                 self.cpus)
+#         elif self.recordFormat == "mp4":
+#             self.quality_video = "x264enc qp-min=5 qp-max=5 speed-preset=1 threads={0} ! h264parse ! video/x-h264, profile=baseline".format(
+#                 self.cpus)
+#     return self.quality_video
 
 
-def on__formats_changed(self, *args):
-    format = self.settings.get_enum("video-format")
-    if format == 0:
-        self.recordFormat = "webm"
-    if format == 1:
-        self.recordFormat = "mkv"
-    if format == 2:
-        self.recordFormat = "mp4"
-    return self.recordFormat
+# def on__formats_changed(self, *args):
+#     format = self.settings.get_enum("video-format")
+#     if format == 0:
+#         self.recordFormat = "webm"
+#     if format == 1:
+#         self.recordFormat = "mkv"
+#     if format == 2:
+#         self.recordFormat = "mp4"
+#     return self.recordFormat
 
 
 # def delay_button_change(self, spin):
@@ -90,239 +90,239 @@ def on__formats_changed(self, *args):
 #         self.settings.set_boolean('record-mouse-cursor-switch', False)
 
 
-def on__select_area(self):
-    coordinate = Popen("slop -n -c 0.3,0.4,0.6,0.4 -l -t 0 -f '%w %h %x %y'", shell=True,
-                       stdout=PIPE).communicate()
-    listCoor = [int(i) for i in coordinate[0].decode().split()]
-    if not listCoor[0] or not listCoor[1]:
-        notification = Gio.Notification.new(constants["APPNAME"])
-        notification.set_body(_("Please re-select the area"))
-        self.application.send_notification(None, notification)
-        return
-
-    startx, starty, endx, endy = listCoor[2], listCoor[3], listCoor[2] + listCoor[0] - 1, listCoor[
-        1] + listCoor[3] - 1
-    if listCoor[0] % 2 == 0 and listCoor[1] % 2 == 0:
-        self.widthArea = endx - startx + 1
-        self.heightArea = endy - starty + 1
-    elif listCoor[0] % 2 == 0 and listCoor[1] % 2 == 1:
-        self.widthArea = endx - startx + 1
-        self.heightArea = endy - starty + 2
-    elif listCoor[0] % 2 == 1 and listCoor[1] % 2 == 1:
-        self.widthArea = endx - startx
-        self.heightArea = endy - starty
-    elif listCoor[0] % 2 == 1 and listCoor[1] % 2 == 0:
-        self.widthArea = endx - startx + 2
-        self.heightArea = endy - starty + 1
-
-    self.coordinateArea = "startx={} starty={} endx={} endy={}".format(startx, starty, endx, endy)
-    self.coordinateMode = True
-
-
-def on__select_area_wayland(self):
-    self.waylandcoordinates = self.GNOMESelectArea.call_sync("SelectArea", None, Gio.DBusProxyFlags.NONE, -1, None)
-    self.coordinateMode = True
+# def on__select_area(self):
+#     coordinate = Popen("slop -n -c 0.3,0.4,0.6,0.4 -l -t 0 -f '%w %h %x %y'", shell=True,
+#                        stdout=PIPE).communicate()
+#     listCoor = [int(i) for i in coordinate[0].decode().split()]
+#     if not listCoor[0] or not listCoor[1]:
+#         notification = Gio.Notification.new(constants["APPNAME"])
+#         notification.set_body(_("Please re-select the area"))
+#         self.application.send_notification(None, notification)
+#         return
+#
+#     startx, starty, endx, endy = listCoor[2], listCoor[3], listCoor[2] + listCoor[0] - 1, listCoor[
+#         1] + listCoor[3] - 1
+#     if listCoor[0] % 2 == 0 and listCoor[1] % 2 == 0:
+#         self.widthArea = endx - startx + 1
+#         self.heightArea = endy - starty + 1
+#     elif listCoor[0] % 2 == 0 and listCoor[1] % 2 == 1:
+#         self.widthArea = endx - startx + 1
+#         self.heightArea = endy - starty + 2
+#     elif listCoor[0] % 2 == 1 and listCoor[1] % 2 == 1:
+#         self.widthArea = endx - startx
+#         self.heightArea = endy - starty
+#     elif listCoor[0] % 2 == 1 and listCoor[1] % 2 == 0:
+#         self.widthArea = endx - startx + 2
+#         self.heightArea = endy - starty + 1
+#
+#     self.coordinateArea = "startx={} starty={} endx={} endy={}".format(startx, starty, endx, endy)
+#     self.coordinateMode = True
 
 
-def on__sound_switch(self, *args):
-    if self._sound_on_switch.get_active():
-        self.recordSoundOn = True
-
-        import pulsectl
-        with pulsectl.Pulse() as pulse:
-            self.soundOnSource = pulse.sink_list()[0].name
-            self.settings.set_boolean('record-audio-switch', True)
-            if self.recordFormat == "webm" or self.recordFormat == "mkv":
-                self.soundOn = " pulsesrc provide-clock=false device='{}.monitor' buffer-time=20000000 ! 'audio/x-raw,depth=24,channels=2,rate=44100,format=F32LE,payload=96' ! queue ! audioconvert ! vorbisenc ! queue ! mux.".format(
-                    self.soundOnSource)
-
-            elif self.recordFormat == "mp4":
-                self.soundOn = " pulsesrc buffer-time=20000000 device='{}.monitor' ! 'audio/x-raw,channels=2,rate=48000' ! queue ! audioconvert ! queue ! opusenc bitrate=512000 ! queue ! mux.".format(
-                    self.soundOnSource)
-        return self.soundOn
-    else:
-        self.recordSoundOn = False
-        self.settings.set_boolean('record-audio-switch', False)
+# def on__select_area_wayland(self):
+#     self.waylandcoordinates = self.GNOMESelectArea.call_sync("SelectArea", None, Gio.DBusProxyFlags.NONE, -1, None)
+#     self.coordinateMode = True
 
 
-def start_recording(self, *args):
-    if self.isFullscreenMode:
-        self.coordinateMode = False
-        record(self)
-    elif self.isWindowMode:
-        print('window mode')
-    else:
-        if self.displayServer == "wayland":
-            on__select_area_wayland(self)
-        else:
-            on__select_area(self)
-        record(self)
+# def on__sound_switch(self, *args):
+#     if self._sound_on_computer.get_active():
+#         self.recordSoundOn = True
+#
+#         import pulsectl
+#         with pulsectl.Pulse() as pulse:
+#             self.soundOnSource = pulse.sink_list()[0].name
+#             self.settings.set_boolean('sound-on-computer', True)
+#             if self.recordFormat == "webm" or self.recordFormat == "mkv":
+#                 self.soundOn = " pulsesrc provide-clock=false device='{}.monitor' buffer-time=20000000 ! 'audio/x-raw,depth=24,channels=2,rate=44100,format=F32LE,payload=96' ! queue ! audioconvert ! vorbisenc ! queue ! mux.".format(
+#                     self.soundOnSource)
+#
+#             elif self.recordFormat == "mp4":
+#                 self.soundOn = " pulsesrc buffer-time=20000000 device='{}.monitor' ! 'audio/x-raw,channels=2,rate=48000' ! queue ! audioconvert ! queue ! opusenc bitrate=512000 ! queue ! mux.".format(
+#                     self.soundOnSource)
+#         return self.soundOn
+#     else:
+#         self.recordSoundOn = False
+#         self.settings.set_boolean('sound-on-computer', False)
 
 
-def record(self, *args):
-    if self.delayBeforeRecording > 0:
-        self._main_stack.set_visible_child(self._delay_box)
-        self._record_stop_record_button_stack.set_visible_child(self._cancel_button)
-        self._menu_stack_revealer.set_reveal_child(False)
-        self.isrecordingwithdelay = True
-        delay(self, *args)
-    else:
-        record_logic(self, *args)
+# def start_recording(self, *args):
+#     if self.isFullscreenMode:
+#         self.coordinateMode = False
+#         record(self)
+#     elif self.isWindowMode:
+#         print('window mode')
+#     else:
+#         if self.displayServer == "wayland":
+#             on__select_area_wayland(self)
+#         else:
+#             on__select_area(self)
+#         record(self)
 
 
-def record_logic(self, *args):
-    if self.iscancelled:
-        self._main_stack.set_visible_child(self._main_screen_box)
-        self._record_stop_record_button_stack.set_visible_child(self._record_button)
-        self._menu_stack.set_visible_child(self._menu_button)
-        self.iscancelled = False
-    else:
-        self._record_stop_record_button_stack.set_visible_child(self._stop_record_button)
-        self._main_stack.set_visible_child(self._paused_start_stack_box)
-        self._menu_stack.set_visible_child(self._pause_record_button)
-        self.label_context = self._time_recording_label.get_style_context()
-        self.label_context.add_class("recording")
-
-        self.quality_video = on__quality_changed(self, *args)
-        self.videoFrames = on__frames_changed(self, *args)
-        self.recordFormat = on__formats_changed(self, *args)
-
-        self.soundOn = on__sound_switch(self, *args)
-        fileNameTime = _(constants["APPNAME"]) + "-" + datetime.datetime.now().strftime("%Y-%m-%d-%H:%M:%S")
-        videoFolder = self.settings.get_string('path-to-save-video-folder')
-        self.fileName = os.path.join(videoFolder, fileNameTime)
-
-        if self.recordFormat == "webm":
-            self.mux = "webmmux"
-            self.extension = ".webm"
-
-        elif self.recordFormat == "mkv":
-            self.mux = "matroskamux"
-            self.extension = ".mkv"
-
-        elif self.recordFormat == "mp4":
-            self.mux = "mp4mux"
-            self.extension = ".mp4"
-
-        if self.displayServer == "wayland":
-            RecorderPipeline = "{0} ! queue ! {1}".format(self.quality_video, self.mux)
-            if self.coordinateMode is True:
-                self.GNOMEScreencast.call_sync(
-                    "ScreencastArea",
-                    GLib.Variant.new_tuple(
-                        GLib.Variant("i", self.waylandcoordinates[0]),
-                        GLib.Variant("i", self.waylandcoordinates[1]),
-                        GLib.Variant("i", self.waylandcoordinates[2]),
-                        GLib.Variant("i", self.waylandcoordinates[3]),
-                        GLib.Variant.new_string(self.fileName + self.extension),
-                        GLib.Variant("a{sv}",
-                                     {"framerate": GLib.Variant("i", int(self.videoFrames)),
-                                      "draw-cursor": GLib.Variant("b", self.recordMouse),
-                                      "pipeline": GLib.Variant("s", RecorderPipeline)}
-                                     ),
-                    ),
-                    Gio.DBusProxyFlags.NONE,
-                    -1,
-                    None)
-                self.coordinateMode == False
-            else:
-                self.GNOMEScreencast.call_sync(
-                    "Screencast",
-                    GLib.Variant.new_tuple(
-                        GLib.Variant.new_string(self.fileName + self.extension),
-                        GLib.Variant("a{sv}",
-                                     {"framerate": GLib.Variant("i", int(self.videoFrames)),
-                                      "draw-cursor": GLib.Variant("b", self.recordMouse),
-                                      "pipeline": GLib.Variant("s", RecorderPipeline)}
-                                     ),
-                    ),
-                    Gio.DBusProxyFlags.NONE,
-                    -1,
-                    None)
-        else:
-            if self.coordinateMode is True:
-                video_str = "gst-launch-1.0 --eos-on-shutdown ximagesrc show-pointer={0} " + self.coordinateArea + " ! videoscale ! video/x-raw,width={1},height={2},framerate={3}/1 ! queue ! videoscale ! videoconvert ! {4} ! queue ! {5} name=mux ! queue ! filesink location='{6}'{7}"
-                if self.recordSoundOn is True:
-                    self.video = Popen(
-                        video_str.format(self.recordMouse, self.widthArea, self.heightArea,
-                                         self.videoFrames, self.quality_video, self.mux, self.fileName,
-                                         self.extension) + self.soundOn, shell=True)
-
-                else:
-                    self.video = Popen(
-                        video_str.format(self.recordMouse, self.widthArea, self.heightArea,
-                                         self.videoFrames, self.quality_video, self.mux, self.fileName,
-                                         self.extension), shell=True)
-
-                self.coordinateMode = False
-            else:
-                if self.recordSoundOn is True:
-                    self.video = Popen(
-                        self.video_str.format(self.recordMouse, self.videoFrames, self.quality_video,
-                                              self.mux, self.fileName, self.extension) + self.soundOn,
-                        shell=True)
-                else:
-                    self.video = Popen(
-                        self.video_str.format(self.recordMouse, self.videoFrames, self.quality_video,
-                                              self.mux, self.fileName, self.extension), shell=True)
-
-        self.isrecording = True
-        self.istimerrunning = True
-        self.playsound('/com/github/amikha1lov/RecApp/sounds/chime.ogg')
+# def record(self, *args):
+#     if self.delayBeforeRecording > 0:
+#         self._main_stack.set_visible_child(self._delay_box)
+#         self._record_stop_record_button_stack.set_visible_child(self._cancel_button)
+#         self._menu_stack_revealer.set_reveal_child(False)
+#         self.isrecordingwithdelay = True
+#         delay(self, *args)
+#     else:
+#         record_logic(self, *args)
 
 
-def delay(self, *args):
-    self.time_delay = (self.delayBeforeRecording * 100)
+# def record_logic(self, *args):
+#     if self.iscancelled:
+#         self._main_stack.set_visible_child(self._main_screen_box)
+#         self._record_stop_record_button_stack.set_visible_child(self._record_button)
+#         self._menu_stack.set_visible_child(self._menu_button)
+#         self.iscancelled = False
+#     else:
+#         self._record_stop_record_button_stack.set_visible_child(self._stop_record_button)
+#         self._main_stack.set_visible_child(self._paused_start_stack_box)
+#         self._menu_stack.set_visible_child(self._pause_record_button)
+#         self.label_context = self._time_recording_label.get_style_context()
+#         self.label_context.add_class("recording")
+#
+#         self.quality_video = on__quality_changed(self, *args)
+#         self.videoFrames = on__frames_changed(self, *args)
+#         self.recordFormat = on__formats_changed(self, *args)
+#
+#         self.soundOn = on__sound_switch(self, *args)
+#         fileNameTime = _(constants["APPNAME"]) + "-" + datetime.datetime.now().strftime("%Y-%m-%d-%H:%M:%S")
+#         videoFolder = self.settings.get_string('path-to-save-video-folder')
+#         self.fileName = os.path.join(videoFolder, fileNameTime)
+#
+#         if self.recordFormat == "webm":
+#             self.mux = "webmmux"
+#             self.extension = ".webm"
+#
+#         elif self.recordFormat == "mkv":
+#             self.mux = "matroskamux"
+#             self.extension = ".mkv"
+#
+#         elif self.recordFormat == "mp4":
+#             self.mux = "mp4mux"
+#             self.extension = ".mp4"
+#
+#         if self.displayServer == "wayland":
+#             RecorderPipeline = "{0} ! queue ! {1}".format(self.quality_video, self.mux)
+#             if self.coordinateMode is True:
+#                 self.GNOMEScreencast.call_sync(
+#                     "ScreencastArea",
+#                     GLib.Variant.new_tuple(
+#                         GLib.Variant("i", self.waylandcoordinates[0]),
+#                         GLib.Variant("i", self.waylandcoordinates[1]),
+#                         GLib.Variant("i", self.waylandcoordinates[2]),
+#                         GLib.Variant("i", self.waylandcoordinates[3]),
+#                         GLib.Variant.new_string(self.fileName + self.extension),
+#                         GLib.Variant("a{sv}",
+#                                      {"framerate": GLib.Variant("i", int(self.videoFrames)),
+#                                       "draw-cursor": GLib.Variant("b", self.recordMouse),
+#                                       "pipeline": GLib.Variant("s", RecorderPipeline)}
+#                                      ),
+#                     ),
+#                     Gio.DBusProxyFlags.NONE,
+#                     -1,
+#                     None)
+#                 self.coordinateMode == False
+#             else:
+#                 self.GNOMEScreencast.call_sync(
+#                     "Screencast",
+#                     GLib.Variant.new_tuple(
+#                         GLib.Variant.new_string(self.fileName + self.extension),
+#                         GLib.Variant("a{sv}",
+#                                      {"framerate": GLib.Variant("i", int(self.videoFrames)),
+#                                       "draw-cursor": GLib.Variant("b", self.recordMouse),
+#                                       "pipeline": GLib.Variant("s", RecorderPipeline)}
+#                                      ),
+#                     ),
+#                     Gio.DBusProxyFlags.NONE,
+#                     -1,
+#                     None)
+#         else:
+#             if self.coordinateMode is True:
+#                 video_str = "gst-launch-1.0 --eos-on-shutdown ximagesrc show-pointer={0} " + self.coordinateArea + " ! videoscale ! video/x-raw,width={1},height={2},framerate={3}/1 ! queue ! videoscale ! videoconvert ! {4} ! queue ! {5} name=mux ! queue ! filesink location='{6}'{7}"
+#                 if self.recordSoundOn is True:
+#                     self.video = Popen(
+#                         video_str.format(self.recordMouse, self.widthArea, self.heightArea,
+#                                          self.videoFrames, self.quality_video, self.mux, self.fileName,
+#                                          self.extension) + self.soundOn, shell=True)
+#
+#                 else:
+#                     self.video = Popen(
+#                         video_str.format(self.recordMouse, self.widthArea, self.heightArea,
+#                                          self.videoFrames, self.quality_video, self.mux, self.fileName,
+#                                          self.extension), shell=True)
+#
+#                 self.coordinateMode = False
+#             else:
+#                 if self.recordSoundOn is True:
+#                     self.video = Popen(
+#                         self.video_str.format(self.recordMouse, self.videoFrames, self.quality_video,
+#                                               self.mux, self.fileName, self.extension) + self.soundOn,
+#                         shell=True)
+#                 else:
+#                     self.video = Popen(
+#                         self.video_str.format(self.recordMouse, self.videoFrames, self.quality_video,
+#                                               self.mux, self.fileName, self.extension), shell=True)
+#
+#         self.isrecording = True
+#         self.istimerrunning = True
+#         self.playsound('/com/github/amikha1lov/RecApp/sounds/chime.ogg')
 
-    def countdown(*args):
-        if self.time_delay > 0:
-            self.time_delay -= 10
-            GLib.timeout_add(100, countdown)
-            self._delay_label.set_label(str((self.time_delay // 100) + 1))
-        else:
-            self.isrecordingwithdelay = False
-            self._menu_stack_revealer.set_reveal_child(True)
-            record_logic(self, *args)
-            self.time_delay = (self.delayBeforeRecording * 100)
 
-    countdown(*args)
+# def delay(self, *args):
+#     self.time_delay = (self.delayBeforeRecording * 100)
+#
+#     def countdown(*args):
+#         if self.time_delay > 0:
+#             self.time_delay -= 10
+#             GLib.timeout_add(100, countdown)
+#             self._delay_label.set_label(str((self.time_delay // 100) + 1))
+#         else:
+#             self.isrecordingwithdelay = False
+#             self._menu_stack_revealer.set_reveal_child(True)
+#             record_logic(self, *args)
+#             self.time_delay = (self.delayBeforeRecording * 100)
+#
+#     countdown(*args)
 
 
-def cancel_delay(self, *args):
-    self.time_delay = 0
-    self.iscancelled = True
+# def cancel_delay(self, *args):
+#     self.time_delay = 0
+#     self.iscancelled = True
 
 
-def stop_recording(self, *args):
-    if self.displayServer == "wayland":
-        self.GNOMEScreencast.call_sync(
-            "StopScreencast",
-            None,
-            Gio.DBusCallFlags.NONE,
-            -1,
-            None)
-
-    else:
-        self.video.send_signal(signal.SIGINT)
-
-    notification = Gio.Notification.new(constants["APPNAME"])
-    notification.set_body(_("Recording is complete!"))
-    notification.add_button(_("Open Folder"), "app.open-folder")
-    notification.add_button(_("Open File"), "app.open-file")
-    notification.set_default_action("app.open-file")
-    self.application.send_notification(None, notification)
-
-    self.isrecording = False
-    self.istimerrunning = False
-
-    self._record_stop_record_button_stack.set_visible_child(self._record_button)
-    self._paused_start_stack.set_visible_child(self._recording_label)
-    self._main_stack.set_visible_child(self._main_screen_box)
-    self._menu_stack.set_visible_child(self._menu_button)
-    self.label_context.remove_class("recording")
-
-    self.elapsed_time = datetime.timedelta()
-    self._time_recording_label.set_label(str(self.elapsed_time).replace(":", "∶"))
+# def stop_recording(self, *args):
+#     if self.displayServer == "wayland":
+#         self.GNOMEScreencast.call_sync(
+#             "StopScreencast",
+#             None,
+#             Gio.DBusCallFlags.NONE,
+#             -1,
+#             None)
+#
+#     else:
+#         self.video.send_signal(signal.SIGINT)
+#
+#     notification = Gio.Notification.new(constants["APPNAME"])
+#     notification.set_body(_("Recording is complete!"))
+#     notification.add_button(_("Open Folder"), "app.open-folder")
+#     notification.add_button(_("Open File"), "app.open-file")
+#     notification.set_default_action("app.open-file")
+#     self.application.send_notification(None, notification)
+#
+#     self.isrecording = False
+#     self.istimerrunning = False
+#
+#     self._record_stop_record_button_stack.set_visible_child(self._record_button)
+#     self._paused_start_stack.set_visible_child(self._recording_label)
+#     self._main_stack.set_visible_child(self._main_screen_box)
+#     self._menu_stack.set_visible_child(self._menu_button)
+#     self.label_context.remove_class("recording")
+#
+#     self.elapsed_time = datetime.timedelta()
+#     self._time_recording_label.set_label(str(self.elapsed_time).replace(":", "∶"))
 
 
 # def delete_event(self, w, h):
@@ -330,7 +330,7 @@ def stop_recording(self, *args):
 #         stop_recording(self)
 
 
-def quit_app(self, *args):
-    if self.isrecording:
-        stop_recording(self)
-    self.destroy()
+# def quit_app(self, *args):
+#     if self.isrecording:
+#         stop_recording(self)
+#     self.destroy()
